Data model first draft.

+++ MarvelCodingExercise/Model/OnDemand/MarvelResponse.swift
@@ -18,7 +18,7 @@ struct MarvelResponse: Decodable {
         let count: Int?
         let results: [Character]?
         struct Character: Decodable {
-            let id: Int?
+            let id: Int64?
             let name: String?
             let description: String?
             let modified: String?
